pyproject.toml
[tool.poetry]
name = "ID"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "2.3.0"
pytest = "7.4.3"
pydantic = "2.7.3"
fastapi = "0.111.0"
uvicorn = "0.25.0"
python-multipart = "0.0.9"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pyjwt = "^2.8.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
